Move Worker and SlaveWorker to App\Slave
