+++ src/Slave/Command/ProbeDispatcherCommand.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Slave\Command;
 
-use App\DependencyInjection\WorkerManager;
 use App\Slave\Configuration;
 use App\Slave\Instruction;
 use App\Slave\Process\SymfonyProcessFactory;
@@ -12,6 +11,7 @@ use App\Slave\Task\FetchConfiguration;
 use App\Slave\Task\PublishStatistics;
 use App\Slave\Worker\Queue;
 use App\Slave\Worker\StatsManager;
+use App\Slave\Worker\WorkerManager;
 use App\Version\GitVersionReader;
 use Exception;
 use Psr\Log\LoggerInterface;
